crypto: mvebu_cesa: fix some build warnings
Reviewed-on: http://vgitil04.il.marvell.com:8080/6545
Tested-by: Star_Automation <star@marvell.com>

+++ drivers/crypto/mvebu_cesa/cesa_dev.c
@@ -142,7 +142,7 @@ int run_cesa_debug(CESA_DEBUG *cesa_debug)
 }
 
 
-static int
+static long
 cesadev_ioctl(
 	struct file *filp,
 	unsigned int cmd,
